chatbot/requirements.txt
Flask
weaviate-client
pandas
tqdm
python-dotenv
flask-cors